backend/requirements.txt
fastapi
pydantic
pytesseract
streamlit
Pillow
uvicorn
matplotlib
requests